backend/src/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
neo4j==5.20.0
langchain==0.1.16
openai==1.30.1
langchain-experimental==0.0.57